+++ CHANGELOG.rst
@@ -26,6 +26,7 @@ Added
 
 * Added ``Reporter`` and ``JSONReporter`` classes to be able to expose execution results
 * Add new config option ``report_name`` to the available settings
+* New exception type ``PreconditionFailedError`` indicating that the precondition failed and no need to raise an error
 
 Changed
 ~~~~~~~
@@ -40,6 +41,9 @@ Changed
 * Pillar prepare its Reporter for report generation
 * Pillar has a new argument to set the pillar's reporter easily
 * CLI's enforce command now calls the Pillar's prepared Reporter to do the report
+* "No changes made by" messages now info logs instead of warnings
+* Commit changes only if the Law has passing rules
+* If ``PreconditionFailedError`` raised, do not log error messages, log a warning instead
 
 Fixed
 ~~~~~